Add OperationCenter for multi-operation progress tracking and cancellation
Extend DownloadCenter into a centralized OperationCenter that tracks all
long-running tasks (downloads, BAM/VCF imports, bundle builds) with per-
operation progress, bundle locking to prevent concurrent mutations, and
user-cancellable operations. Adds Operations menu with live status and a
floating Operations panel (Shift-Option-Cmd-O). Migrates BAM and VCF
import paths from single ActivityIndicatorView to OperationCenter. Also
includes strand color toggle for read rendering and samtools consensus
-a flag fix.

+++ Sources/LungfishIO/Formats/SAM/SAMParser.swift
@@ -346,7 +346,7 @@ public enum SAMParser {
         if qualString == "*" {
             qualities = []
         } else {
-            qualities = qualString.utf8.map { max(0, $0 - 33) }
+            qualities = qualString.utf8.map { UInt8(max(0, Int($0) - 33)) }
         }
 
         // Optional tags (fields 11+)
